Headhunter has successfully been dockerized.
There used to be many open "secrets' in the application files. A number of environment variables have been introduced to mitigate that.

Many small changes have been made to make the dockerization happen. Since this has been a process of trial-and-error, there are few commits.
