--- src/main/resources/application-docker.yml
@@ -1,29 +1,0 @@
-active-application-yml: DOCKER in main
-
-spring:
-  datasource:
-    url: jdbc:mysql://mysql:3306/headhunter-prod?createDatabaseIfNotExist=true&autoReconnect=true&useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=Europe/Berlin
-    username: ${MYSQL_PASSWORD}
-    password: ${MYSQL_USERNAME}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: validate
-    generate-ddl: true
-    show-sql: true
-
-api:
-  endpoint:
-    base-url-account: /api/v1/account
-    base-url-job: /api/v1/job
-    base-url-ad: /api/v1/ad
-    base-url-accountInfo: /api/v1/accountInfo
-
-ai:
-  openai:
-    endpoint: ${OPENAI_API_ENDPOINT}
-    api-key: ${OPENAI_API_KEY}
-
-logging:
-  level:
-    org.hibernate: DEBUG
