+++ src/main/resources/application-prod.yml
@@ -1,0 +1,29 @@
+active-application-yml: PRODUCTION in main
+
+spring:
+  config:
+    activate:
+      on-profile: prod
+  #  config:
+  #    import: "optional:configtree:/run/secrets/"
+  datasource:
+    url: jdbc:${DB_NAME}://${DB_PROD_HOSTNAME}:${DB_PORT}/${DB_PRODUCTION_DATABASE_NAME}
+    username: ${DB_USER_USERNAME}
+    password: ${DB_USER_PASSWORD}
+  jpa:
+    hibernate:
+      ddl-auto: validate
+    generate-ddl: true
+    show-sql: true
+
+api:
+  endpoint:
+    base-url-account: /api/v1/account
+    base-url-job: /api/v1/job
+    base-url-ad: /api/v1/ad
+    base-url-accountInfo: /api/v1/accountInfo
+
+ai:
+  openai:
+    endpoint: ${OPENAI_API_ENDPOINT}
+    api-key: ${OPENAI_API_KEY}
