+++ src/main/resources/application.yml
@@ -1,14 +1,13 @@
-active-application-yml: Default in main
+active-application-yml: DEV in main
 
 spring:
   datasource:
-    url: jdbc:mysql://localhost:3306/headhunter-prod?createDatabaseIfNotExist=true&autoReconnect=true&useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=Europe/Berlin
-    username: ${MYSQL_USER_USERNAME}
-    password: ${MYSQL_USER_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:${DB_NAME}://${DB_DEV_HOSTNAME}:${DB_PORT}/${DB_DEV_DATABASE_NAME}?createDatabaseIfNotExist=true&autoReconnect=true&useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=Europe/Berlin
+    username: ${DB_ADMIN_USERNAME}
+    password: ${DB_ADMIN_PASSWORD}
   jpa:
     hibernate:
-      ddl-auto: validate
+      ddl-auto: create
     generate-ddl: true
     show-sql: true
 
@@ -21,8 +20,6 @@ api:
 
 ai:
   openai:
-    endpoint: https://api.openai.com/v1/chat/completions
-    api-key: hJe8Atw0eTUIumr5ZxaST3BlbkFJwEzMxJNv73Ok5HSk2xAc
-#    endpoint: ${OPENAI_API_ENDPOINT}
-#    api-key: ${OPENAI_API_KEY}
+    endpoint: ${OPENAI_API_ENDPOINT}
+    api-key: ${OPENAI_API_KEY}
