--- src/test/java/se/sprinta/headhunterbackend/account/AccountServiceComplmentaryH2Test.java
@@ -1,42 +1,0 @@
-package se.sprinta.headhunterbackend.account;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import se.sprinta.headhunterbackend.H2DatabaseInitializer;
-import se.sprinta.headhunterbackend.system.exception.ObjectNotFoundException;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-@SpringBootTest
-@ActiveProfiles("test-h2")
-public class AccountServiceComplmentaryH2Test {
-
-    @Autowired
-    private AccountService accountService;
-
-    @Autowired
-    private H2DatabaseInitializer h2DbInit;
-
-    @BeforeEach
-    void setUp() {
-        this.h2DbInit.initializeH2Database();
-    }
-
-    @AfterEach
-    void tearDown() {
-        this.h2DbInit.clearH2Database();
-    }
-
-    @Test
-    @DisplayName("delete - Success")
-    void test_Delete_Success() {
-        this.accountService.delete("user1-h2@hh.se");
-
-        assertThrows(ObjectNotFoundException.class, () -> this.accountService.findById("user1-h2@hh.se"));
-    }
-}
