--- src/test/resources/application-test-h2.yml
@@ -1,24 +1,0 @@
-active-application-yml: H2 in test
-
-spring:
-  datasource:
-    profiles:
-      active: test-h2
-    url: jdbc:h2:mem:headhunter-test-h2
-    driverClassName: org.h2.Driver
-    username: ${H2_USERNAME}
-    password: ${H2_PASSWORD}
-  jpa:
-    database-platform: org.hibernate.dialect.H2Dialect
-    hibernate:
-      ddl-auto: create
-    properties:
-      format_sql: true
-      use_sql_comments: true
-      generate_statistics: true
-    show-sql: true
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-
