--- src/test/resources/application-test-mysql.yml
@@ -1,31 +1,0 @@
-active-application-yml: MySQL in test
-
-spring:
-  datasource:
-    profiles:
-      active: test-mysql
-    url: jdbc:mysql://localhost:3306/headhunter-test?createDatabaseIfNotExist=true&autoReconnect=true&useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=Europe/Berlin
-    username: ${MYSQL_ADMIN_USERNAME}
-    password: ${MYSQL_ADMIN_PASSWORD}
-  jpa:
-    hibernate:
-      ddl-auto: create
-    generate-ddl: true
-    show-sql: true
-
-api:
-  endpoint:
-    base-url-account: /api/v1/account
-    base-url-job: /api/v1/job
-    base-url-ad: /api/v1/ad
-    base-url-accountInfo: /api/v1/accountInfo
-
-ai:
-  openai:
-    endpoint: ${OPENAI_API_ENDPOINT}
-    api-key: ${OPENAI_API_KEY}
-# ai:
-#   openai:
-#     endpoint: https://api.openai.com/v1/chat/completions
-#     api-key: hJe8Atw0eTUIumr5ZxaST3BlbkFJwEzMxJNv73Ok5HSk2xAc
-
